[API stabilization] ProtoBuf rework
    * Transition the the sealed class + companion + builder scheme
    * ProtoId renamed to ProtoNumber and ProtoNumberType to ProtoIntegerType to be better aligned with ProtoBuf specification
    * All internal implementation details are hidden from the public package
    * Hide ProtobufDecodingException from public API
