+++ formats/protobuf/commonMain/src/kotlinx/serialization/protobuf/internal/ProtobufReader.kt
@@ -38,10 +38,10 @@ internal class ProtobufReader(private val input: ByteArrayInput) {
 
     fun skipElement() {
         when (currentType) {
-            ProtoBuf.VARINT -> readInt(ProtoNumberType.DEFAULT)
-            ProtoBuf.i64 -> readLong(ProtoNumberType.FIXED)
-            ProtoBuf.SIZE_DELIMITED -> readByteArray()
-            ProtoBuf.i32 -> readInt(ProtoNumberType.FIXED)
+            VARINT -> readInt(ProtoIntegerType.DEFAULT)
+            i64 -> readLong(ProtoIntegerType.FIXED)
+            SIZE_DELIMITED -> readByteArray()
+            i32 -> readInt(ProtoIntegerType.FIXED)
             else -> throw ProtobufDecodingException("Unsupported start group or end group wire type: $currentType")
         }
     }
@@ -52,7 +52,7 @@ internal class ProtobufReader(private val input: ByteArrayInput) {
     }
 
     fun readByteArray(): ByteArray {
-        assertWireType(ProtoBuf.SIZE_DELIMITED)
+        assertWireType(SIZE_DELIMITED)
         return readByteArrayNoTag()
     }
 
@@ -63,7 +63,7 @@ internal class ProtobufReader(private val input: ByteArrayInput) {
     }
 
     fun objectInput(): ByteArrayInput {
-        assertWireType(ProtoBuf.SIZE_DELIMITED)
+        assertWireType(SIZE_DELIMITED)
         return objectTaglessInput()
     }
 
@@ -73,24 +73,24 @@ internal class ProtobufReader(private val input: ByteArrayInput) {
         return input.slice(length)
     }
 
-    fun readInt(format: ProtoNumberType): Int {
-        val wireType = if (format == ProtoNumberType.FIXED) ProtoBuf.i32 else ProtoBuf.VARINT
+    fun readInt(format: ProtoIntegerType): Int {
+        val wireType = if (format == ProtoIntegerType.FIXED) i32 else VARINT
         assertWireType(wireType)
         return decode32(format)
     }
 
     fun readInt32NoTag(): Int = decode32()
 
-    fun readLong(format: ProtoNumberType): Long {
-        val wireType = if (format == ProtoNumberType.FIXED) ProtoBuf.i64 else ProtoBuf.VARINT
+    fun readLong(format: ProtoIntegerType): Long {
+        val wireType = if (format == ProtoIntegerType.FIXED) i64 else VARINT
         assertWireType(wireType)
         return decode64(format)
     }
 
-    fun readLongNoTag(): Long = decode64(ProtoNumberType.DEFAULT)
+    fun readLongNoTag(): Long = decode64(ProtoIntegerType.DEFAULT)
 
     fun readFloat(): Float {
-        assertWireType(ProtoBuf.i32)
+        assertWireType(i32)
         return Float.fromBits(readIntLittleEndian())
     }
 
@@ -117,7 +117,7 @@ internal class ProtobufReader(private val input: ByteArrayInput) {
     }
 
     fun readDouble(): Double {
-        assertWireType(ProtoBuf.i64)
+        assertWireType(i64)
         return Double.fromBits(readLongLittleEndian())
     }
 
@@ -126,7 +126,7 @@ internal class ProtobufReader(private val input: ByteArrayInput) {
     }
 
     fun readString(): String {
-        assertWireType(ProtoBuf.SIZE_DELIMITED)
+        assertWireType(SIZE_DELIMITED)
         val length = decode32()
         checkLength(length)
         return input.readString(length)
@@ -144,20 +144,20 @@ internal class ProtobufReader(private val input: ByteArrayInput) {
         }
     }
 
-    private fun decode32(format: ProtoNumberType = ProtoNumberType.DEFAULT): Int = when (format) {
-        ProtoNumberType.DEFAULT -> input.readVarint64(false).toInt()
-        ProtoNumberType.SIGNED -> decodeSignedVarintInt(
+    private fun decode32(format: ProtoIntegerType = ProtoIntegerType.DEFAULT): Int = when (format) {
+        ProtoIntegerType.DEFAULT -> input.readVarint64(false).toInt()
+        ProtoIntegerType.SIGNED -> decodeSignedVarintInt(
             input
         )
-        ProtoNumberType.FIXED -> readIntLittleEndian()
+        ProtoIntegerType.FIXED -> readIntLittleEndian()
     }
 
-    private fun decode64(format: ProtoNumberType = ProtoNumberType.DEFAULT): Long = when (format) {
-        ProtoNumberType.DEFAULT -> input.readVarint64(false)
-        ProtoNumberType.SIGNED -> decodeSignedVarintLong(
+    private fun decode64(format: ProtoIntegerType = ProtoIntegerType.DEFAULT): Long = when (format) {
+        ProtoIntegerType.DEFAULT -> input.readVarint64(false)
+        ProtoIntegerType.SIGNED -> decodeSignedVarintLong(
             input
         )
-        ProtoNumberType.FIXED -> readLongLittleEndian()
+        ProtoIntegerType.FIXED -> readLongLittleEndian()
     }
 
     /**
