+++ formats/protobuf/commonTest/src/kotlinx/serialization/PolymorphismTestData.kt
@@ -8,7 +8,7 @@ import kotlinx.serialization.modules.*
 import kotlinx.serialization.protobuf.*
 
 @Serializable
-open class PolyBase(@ProtoId(1) val id: Int) {
+open class PolyBase(@ProtoNumber(1) val id: Int) {
     override fun hashCode(): Int {
         return id
     }
@@ -31,7 +31,7 @@ open class PolyBase(@ProtoId(1) val id: Int) {
 }
 
 @Serializable
-data class PolyDerived(@ProtoId(2) val s: String) : PolyBase(1)
+data class PolyDerived(@ProtoNumber(2) val s: String) : PolyBase(1)
 
 @Serializable
 abstract class SimpleAbstract
