+++ formats/protobuf/commonTest/src/kotlinx/serialization/protobuf/CustomSerializersProtobufTest.kt
@@ -10,8 +10,9 @@ import kotlin.test.*
 
 class CustomSerializersProtobufTest {
 
-    private fun protoBufWithB() =
-        ProtoBuf(serializersModule = serializersModuleOf(B::class, BSerializer))
+    private fun protoBufWithB() = ProtoBuf {
+        serializersModule = serializersModuleOf(B::class, BSerializer)
+    }
 
     @Test
     fun writeCustom() {
@@ -48,39 +49,35 @@ class CustomSerializersProtobufTest {
     @Test
     fun writeCustomInvertedOrder() {
         val obj = C(1, 2)
-        val j = ProtoBuf()
-        val s = j.encodeToHexString(obj).toUpperCase()
+        val s = ProtoBuf.encodeToHexString(obj).toUpperCase()
         assertEquals("10020801", s)
     }
 
     @Test
     fun readCustomInvertedOrder() {
         val obj = C(1, 2)
-        val j = ProtoBuf()
-        val s = j.decodeFromHexString<C>("10020801")
+        val s = ProtoBuf.decodeFromHexString<C>("10020801")
         assertEquals(obj, s)
     }
 
     @Test
     fun writeCustomOmitDefault() {
         val obj = C(b = 2)
-        val j = ProtoBuf()
-        val s = j.encodeToHexString(obj).toUpperCase()
+        val s = ProtoBuf.encodeToHexString(obj).toUpperCase()
         assertEquals("1002", s)
     }
 
     @Test
     fun readCustomOmitDefault() {
         val obj = C(b = 2)
-        val j = ProtoBuf()
-        val s = j.decodeFromHexString<C>("1002")
+        val s = ProtoBuf.decodeFromHexString<C>("1002")
         assertEquals(obj, s)
     }
 
     @Test
     fun writeOptionalList1() {
         val obj = CList1(listOf(C(a = 1), C(b = 2), C(3, 4)))
-        val s = ProtoBuf().encodeToHexString(obj).toUpperCase()
+        val s = ProtoBuf.encodeToHexString(obj).toUpperCase()
         assertEquals("0A04102A08010A0210020A0410040803", s)
     }
 
@@ -88,13 +85,13 @@ class CustomSerializersProtobufTest {
     fun readOptionalList1() {
         val obj = CList1(listOf(C(a = 1), C(b = 2), C(3, 4)))
         val j = "0A04102A08010A0210020A0410040803"
-        assertEquals(obj, ProtoBuf().decodeFromHexString(j))
+        assertEquals(obj, ProtoBuf.decodeFromHexString(j))
     }
 
     @Test
     fun writeOptionalList2a() {
         val obj = CList2(7, listOf(C(a = 5), C(b = 6), C(7, 8)))
-        val s = ProtoBuf().encodeToHexString(obj).toUpperCase()
+        val s = ProtoBuf.encodeToHexString(obj).toUpperCase()
         assertEquals("1204102A0805120210061204100808070807", s)
     }
 
@@ -102,13 +99,13 @@ class CustomSerializersProtobufTest {
     fun readOptionalList2a() {
         val obj = CList2(7, listOf(C(a = 5), C(b = 6), C(7, 8)))
         val j = "08071204102A080512021006120410080807"
-        assertEquals(obj, ProtoBuf().decodeFromHexString<CList2>(j))
+        assertEquals(obj, ProtoBuf.decodeFromHexString<CList2>(j))
     }
 
     @Test
     fun writeOptionalList2b() {
         val obj = CList2(c = listOf(C(a = 5), C(b = 6), C(7, 8)))
-        val s = ProtoBuf().encodeToHexString(obj).toUpperCase()
+        val s = ProtoBuf.encodeToHexString(obj).toUpperCase()
         assertEquals("1204102A080512021006120410080807", s)
     }
 
@@ -116,13 +113,13 @@ class CustomSerializersProtobufTest {
     fun readOptionalList2b() {
         val obj = CList2(c = listOf(C(a = 5), C(b = 6), C(7, 8)))
         val j = "1204102A080512021006120410080807"
-        assertEquals(obj, ProtoBuf().decodeFromHexString<CList2>(j))
+        assertEquals(obj, ProtoBuf.decodeFromHexString<CList2>(j))
     }
 
     @Test
     fun writeOptionalList3a() {
         val obj = CList3(listOf(C(a = 1), C(b = 2), C(3, 4)), 99)
-        val s = ProtoBuf().encodeToHexString(obj).toUpperCase()
+        val s = ProtoBuf.encodeToHexString(obj).toUpperCase()
         assertEquals("0A04102A08010A0210020A04100408031063", s)
     }
 
@@ -130,13 +127,13 @@ class CustomSerializersProtobufTest {
     fun readOptionalList3a() {
         val obj = CList3(listOf(C(a = 1), C(b = 2), C(3, 4)), 99)
         val j = "10630A04102A08010A0210020A0410040803"
-        assertEquals(obj, ProtoBuf().decodeFromHexString<CList3>(j))
+        assertEquals(obj, ProtoBuf.decodeFromHexString<CList3>(j))
     }
 
     @Test
     fun writeOptionalList3b() {
         val obj = CList3(f = 99)
-        val s = ProtoBuf().encodeToHexString(obj).toUpperCase()
+        val s = ProtoBuf.encodeToHexString(obj).toUpperCase()
         assertEquals("1063", s)
     }
 
@@ -144,13 +141,13 @@ class CustomSerializersProtobufTest {
     fun readOptionalList3b() {
         val obj = CList3(f = 99)
         val j = "1063"
-        assertEquals(obj, ProtoBuf().decodeFromHexString<CList3>(j))
+        assertEquals(obj, ProtoBuf.decodeFromHexString<CList3>(j))
     }
 
     @Test
     fun writeOptionalList4a() {
         val obj = CList4(listOf(C(a = 1), C(b = 2), C(3, 4)), 54)
-        val s = ProtoBuf().encodeToHexString(obj).toUpperCase()
+        val s = ProtoBuf.encodeToHexString(obj).toUpperCase()
         assertEquals("10360A04102A08010A0210020A0410040803", s)
     }
 
@@ -158,14 +155,14 @@ class CustomSerializersProtobufTest {
     fun readOptionalList4a() {
         val obj = CList4(listOf(C(a = 1), C(b = 2), C(3, 4)), 54)
         val j = "10360A04102A08010A0210020A0410040803"
-        assertEquals(obj, ProtoBuf().decodeFromHexString<CList4>(j))
+        assertEquals(obj, ProtoBuf.decodeFromHexString<CList4>(j))
     }
 
     @Test
     fun writeOptionalList4b() {
         val obj = CList4(h = 97)
         val j = "1061"
-        val s = ProtoBuf().encodeToHexString(obj).toUpperCase()
+        val s = ProtoBuf.encodeToHexString(obj).toUpperCase()
         assertEquals(j, s)
     }
 
@@ -173,13 +170,13 @@ class CustomSerializersProtobufTest {
     fun readOptionalList4b() {
         val obj = CList4(h = 97)
         val j = "1061"
-        assertEquals(obj, ProtoBuf().decodeFromHexString<CList4>(j))
+        assertEquals(obj, ProtoBuf.decodeFromHexString<CList4>(j))
     }
 
     @Test
     fun writeOptionalList5a() {
         val obj = CList5(listOf(9, 8, 7, 6, 5), 5)
-        val s = ProtoBuf().encodeToHexString(obj).toUpperCase()
+        val s = ProtoBuf.encodeToHexString(obj).toUpperCase()
         assertEquals("100508090808080708060805", s)
     }
 
@@ -187,13 +184,13 @@ class CustomSerializersProtobufTest {
     fun readOptionalList5a() {
         val obj = CList5(listOf(9, 8, 7, 6, 5), 5)
         val j = "100508090808080708060805"
-        assertEquals(obj, ProtoBuf().decodeFromHexString<CList5>(j))
+        assertEquals(obj, ProtoBuf.decodeFromHexString<CList5>(j))
     }
 
     @Test
     fun writeOptionalList5b() {
         val obj = CList5(h = 999)
-        val s = ProtoBuf().encodeToHexString(obj).toUpperCase()
+        val s = ProtoBuf.encodeToHexString(obj).toUpperCase()
         assertEquals("10E707", s)
     }
 
@@ -201,6 +198,6 @@ class CustomSerializersProtobufTest {
     fun readOptionalList5b() {
         val obj = CList5(h = 999)
         val j = "10E707"
-        assertEquals(obj, ProtoBuf().decodeFromHexString<CList5>(j))
+        assertEquals(obj, ProtoBuf.decodeFromHexString<CList5>(j))
     }
 }
