+++ formats/protobuf/jvmMain/src/kotlinx/serialization/protobuf/internal/Bytes.kt
@@ -2,7 +2,7 @@
  * Copyright 2017-2020 JetBrains s.r.o. Use of this source code is governed by the Apache 2.0 license.
  */
 
-package kotlinx.serialization.protobuf
+package kotlinx.serialization.protobuf.internal
 
 internal actual fun Int.reverseBytes(): Int = Integer.reverseBytes(this)
