+++ formats/protobuf/nativeMain/src/kotlinx/serialization/protobuf/internal/Bytes.kt
@@ -2,7 +2,7 @@
  * Copyright 2017-2020 JetBrains s.r.o. Use of this source code is governed by the Apache 2.0 license.
  */
 
-package kotlinx.serialization.protobuf
+package kotlinx.serialization.protobuf.internal
 
 private fun Short.reverseBytes(): Short = (((this.toInt() and 0xff) shl 8) or ((this.toInt() and 0xffff) ushr 8)).toShort()
