+++ runtime/commonMain/src/kotlinx/serialization/descriptors/SerialDescriptors.kt
@@ -213,7 +213,7 @@ public class ClassSerialDescriptorBuilder internal constructor(
      * ```
      * class Data(
      *     val intField: Int? = null, // Optional, has default value
-     *     @ProtoId(1) val longField: Long
+     *     @ProtoNumber(1) val longField: Long
      * )
      *
      * // Corresponding descriptor
